refactor: replace internals package references with utils package in various files
