+++ internal/utils/etcd.go
@@ -1,6 +1,3 @@
-/*
-Copyright © 2024 Beingana Jim Junior and Contributors
-*/
 package utils
 
 import (
