+++ internal/utils/validator.go
@@ -1,4 +1,4 @@
-package helpers
+package utils
 
 import (
 	"encoding/json"
